// task-management-express/src/enums/subscription.ts
export enum TSubscription { //-- task-management
    none = 'none', // This can not be created through admin dashboard
    individual = 'individual', // This can not be created through admin dashboard // we assign standard price Id here 
    business_starter = 'business_starter',
    business_level1 = 'business_level1',
    business_level2 = 'business_level2'
}